refactor: packet
