+++ src/core/packet/proto/message/element.ts
