+++ src/core/packet/proto/oidb/Oidb.0x8FC_2.ts
