+++ src/core/packet/proto/old/Message.ts
