+++ src/onebot/action/extends/UploadForwardMsg.ts
@@ -1,7 +1,7 @@
 import BaseAction from '../BaseAction';
 import { ActionName } from '../types';
 import { FromSchema, JSONSchema } from 'json-schema-to-ts';
-import {PacketMsgTextElement} from "@/core/helper/packet/msg/element";
+import {PacketMsgTextElement} from "@/core/packet/msg/element";
 import {SendTextElement} from "@/core";
